Add matrix class
